+++ tests/NgrokCommandTest.php
@@ -2,13 +2,9 @@
 
 namespace JnJairo\Laravel\Ngrok\Tests;
 
-use Illuminate\Console\OutputStyle;
-use Illuminate\Contracts\Config\Repository;
-use Illuminate\Contracts\Container\Container;
-use JnJairo\Laravel\Ngrok\NgrokCommand;
 use JnJairo\Laravel\Ngrok\NgrokProcessBuilder;
 use JnJairo\Laravel\Ngrok\NgrokWebService;
-use JnJairo\Laravel\Ngrok\Tests\TestCase;
+use JnJairo\Laravel\Ngrok\Tests\OrchestraTestCase as TestCase;
 use Symfony\Component\Process\Process;
 
 /**
@@ -18,18 +14,10 @@ class NgrokCommandTest extends TestCase
 {
     public function test_handle() : void
     {
-        $input = new \Symfony\Component\Console\Input\ArrayInput(['host' => 'example.com', '--port' => '80']);
-        $output = new \Symfony\Component\Console\Output\NullOutput;
+        $host = 'example.com';
+        $port = '80';
 
-        $config = $this->prophesize(Repository::class);
-        $config->get('app.url')->willReturn('http://example.com')->shouldNotBeCalled();
-
-        $container = $this->prophesize(Container::class);
-        $container->make(OutputStyle::class, \Prophecy\Argument::any())->willReturn(
-            new OutputStyle($input, $output)
-        );
-        $container->call(\Prophecy\Argument::any())->shouldBeCalled();
-        $container->make('config')->willReturn($config->reveal())->shouldNotBeCalled();
+        config(['app.url' => '']);
 
         $tunnels = [
             [
@@ -65,30 +53,23 @@ class NgrokCommandTest extends TestCase
         $process->getExitCode()->willReturn(0)->shouldBeCalled();
 
         $processBuilder = $this->prophesize(NgrokProcessBuilder::class);
-        $processBuilder->buildProcess('example.com', '80')->willReturn($process->reveal())->shouldBeCalled();
+        $processBuilder->buildProcess($host, $port)->willReturn($process->reveal())->shouldBeCalled();
 
-        $command = new NgrokCommand($processBuilder->reveal(), $webService->reveal());
-        $command->setLaravel($container->reveal());
-        $command->run($input, $output);
-        $this->assertSame(0, $command->handle());
+        app()->instance(NgrokWebService::class, $webService->reveal());
+        app()->instance(NgrokProcessBuilder::class, $processBuilder->reveal());
 
-        $container->call([$command, 'handle'])->shouldHaveBeenCalled();
+        $this->artisan('ngrok', ['host' => $host, '--port' => $port])
+             ->expectsOutput('Host: ' . $host)
+             ->expectsOutput('Port: ' . $port)
+             ->assertExitCode(0);
     }
 
     public function test_handle_from_config() : void
     {
-        $input = new \Symfony\Component\Console\Input\ArrayInput([]);
-        $output = new \Symfony\Component\Console\Output\NullOutput;
+        $host = 'example.com';
+        $port = '8000';
 
-        $config = $this->prophesize(Repository::class);
-        $config->get('app.url')->willReturn('http://example.com:8000')->shouldBeCalled();
-
-        $container = $this->prophesize(Container::class);
-        $container->make(OutputStyle::class, \Prophecy\Argument::any())->willReturn(
-            new OutputStyle($input, $output)
-        );
-        $container->call(\Prophecy\Argument::any())->shouldBeCalled();
-        $container->make('config')->willReturn($config->reveal())->shouldBeCalled();
+        config(['app.url' => 'http://example.com:8000']);
 
         $tunnels = [
             [
@@ -124,30 +105,23 @@ class NgrokCommandTest extends TestCase
         $process->getExitCode()->willReturn(0)->shouldBeCalled();
 
         $processBuilder = $this->prophesize(NgrokProcessBuilder::class);
-        $processBuilder->buildProcess('example.com', '8000')->willReturn($process->reveal())->shouldBeCalled();
+        $processBuilder->buildProcess($host, $port)->willReturn($process->reveal())->shouldBeCalled();
 
-        $command = new NgrokCommand($processBuilder->reveal(), $webService->reveal());
-        $command->setLaravel($container->reveal());
-        $command->run($input, $output);
-        $this->assertSame(0, $command->handle());
+        app()->instance(NgrokWebService::class, $webService->reveal());
+        app()->instance(NgrokProcessBuilder::class, $processBuilder->reveal());
 
-        $container->call([$command, 'handle'])->shouldHaveBeenCalled();
+        $this->artisan('ngrok')
+             ->expectsOutput('Host: ' . $host)
+             ->expectsOutput('Port: ' . $port)
+             ->assertExitCode(0);
     }
 
     public function test_handle_invalid_host() : void
     {
-        $input = new \Symfony\Component\Console\Input\ArrayInput([]);
-        $output = new \Symfony\Component\Console\Output\NullOutput;
+        $host = 'example.com';
+        $port = '8000';
 
-        $config = $this->prophesize(Repository::class);
-        $config->get('app.url')->willReturn('')->shouldBeCalled();
-
-        $container = $this->prophesize(Container::class);
-        $container->make(OutputStyle::class, \Prophecy\Argument::any())->willReturn(
-            new OutputStyle($input, $output)
-        );
-        $container->call(\Prophecy\Argument::any())->shouldBeCalled();
-        $container->make('config')->willReturn($config->reveal())->shouldBeCalled();
+        config(['app.url' => '']);
 
         $tunnels = [
             [
@@ -170,13 +144,12 @@ class NgrokCommandTest extends TestCase
         $process->getExitCode()->willReturn(0)->shouldNotBeCalled();
 
         $processBuilder = $this->prophesize(NgrokProcessBuilder::class);
-        $processBuilder->buildProcess('example.com', '8000')->willReturn($process->reveal())->shouldNotBeCalled();
+        $processBuilder->buildProcess($host, $port)->willReturn($process->reveal())->shouldNotBeCalled();
 
-        $command = new NgrokCommand($processBuilder->reveal(), $webService->reveal());
-        $command->setLaravel($container->reveal());
-        $command->run($input, $output);
-        $this->assertSame(1, $command->handle());
+        app()->instance(NgrokWebService::class, $webService->reveal());
+        app()->instance(NgrokProcessBuilder::class, $processBuilder->reveal());
 
-        $container->call([$command, 'handle'])->shouldHaveBeenCalled();
+        $this->artisan('ngrok')
+             ->assertExitCode(1);
     }
 }